requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
APScheduler==3.7.0
astroid==2.5.1
async-timeout==3.0.1
asyncpraw==7.2.0
asyncprawcore==2.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
better-profanity==0.7.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
discord-components==1.0.0
discord-ext-menus
discord.py==1.6.0
googletrans
h11==0.9.0
h2==3.2.0
heroku==0.1.4
hpack==3.0.0
hstspreload==2020.12.22
httpcore
httpx
hyperframe==5.2.0
idna==2.10
instant-api-client==0.1.2
isort==5.7.0
lazy-object-proxy==1.5.2
lxml==4.6.3
mccabe==0.6.1
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
Pillow==8.1.2
praw==7.3.0
prawcore==2.2.0
prsaw==0.3.1
psutil==5.8.0
pyfiglet==0.8.post1
pylint==2.7.2
python-dateutil==1.5
pytz==2021.1
randomstuff.py==1.5.1
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.58.0
wrapt==1.12.1
yarl==1.6.3
